+++ CHANGELOG.md
@@ -1,9 +1,37 @@
 # Change Log
 
-All notable changes to the "phone-number-formatter" extension will be documented in this file.
+All notable changes to the "contacts-formatter" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.1.0] - 2025-04-07
+### Added
+- add ```Contact Formatter``` category to commands
+- add Clean List of Names command
 
-- Initial release
+### Changed
+- change to contact-formatter
+
+## [0.0.6] - 2025-04-07
+### Fixed
+- add start number for Generate Name List command
+
+## [0.0.5] - 2025-04-07
+### Added
+- add Generate Name List command
+
+## [0.0.4] - 2025-04-07
+### Added
+- add Clean Phone Number command
+
+## [0.0.3] - 2025-04-07
+### Fixed
+- add +628 starting number
+
+## [0.0.2] - 2025-04-07
+### Fixed
+- check phone number before remove non-numeric characters
+
+## [0.0.1] - 2025-01-29
+### Added
+- Initial release of phone-number-formatter.